requirements.txt
# matplotlib==3.5.1
numpy==1.21.5
opencv_python==4.6.0.66
Pillow==9.5.0
PyQt5==5.15.9
# PySide6==6.5.0
system_hotkey==1.0.3
